added maxHealth to heroes & limits to taking damage &  healing

+++ src/main/java/characters/enemies/Boss.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 public class Boss extends Enemy {
 
     private ArrayList<Item> loot;
-    public Boss(long health, int damage, int xp, int level) {
+    public Boss(int health, int damage, int xp, int level) {
         super(health, damage, xp, level);
         this.loot = new ArrayList<>();
     }
