+++ src/main/java/characters/enemies/Monster.java
@@ -1,7 +1,7 @@
 package characters.enemies;
 
 public class Monster extends Enemy {
-    public Monster(long health, int damage, int xp, int level) {
+    public Monster(int health, int damage, int xp, int level) {
         super(health, damage, xp, level);
     }
 }
